+++ src/Grouping/Base/AbstractList.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace PHPForge\Html\Base;
+namespace PHPForge\Html\Grouping\Base;
 
-use PHPForge\Html\Attribute;
+use PHPForge\Html\Attribute\{Custom\HasAttributes, Custom\HasTagName, Input\HasName};
+use PHPForge\Html\Attribute\{HasClass, HasId, HasLang, HasStyle, HasTabindex, HasTitle};
 use PHPForge\Html\Tag;
-use PHPForge\Widget\Element;
-use PHPForge\Widget\ElementInterface;
+use PHPForge\Widget\{Element, ElementInterface};
 
 use function trim;
 
@@ -16,19 +16,18 @@ use function trim;
  */
 abstract class AbstractList extends Element
 {
-    use Attribute\Custom\HasAttributes;
-    use Attribute\Custom\HasTagName;
-    use Attribute\HasClass;
-    use Attribute\HasId;
-    use Attribute\HasLang;
-    use Attribute\HasStyle;
-    use Attribute\HasTabindex;
-    use Attribute\HasTitle;
-    use Attribute\Input\HasName;
+    use HasAttributes;
+    use HasClass;
+    use HasId;
+    use HasLang;
+    use HasName;
+    use HasStyle;
+    use HasTabindex;
+    use HasTagName;
+    use HasTitle;
 
     protected array $attributes = [];
     protected string $content = '';
-    protected string $tagName = '';
 
     /**
      * Set the `HTML` content value.
