+++ src/Grouping/Ol.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace PHPForge\Html;
+namespace PHPForge\Html\Grouping;
+
+use PHPForge\Html\Attribute\{Input\HasType, Tag\HasReversed, Tag\HasStart};
 
 /**
  * The `<ol> HTML element represents an ordered list of items, typically rendered as a numbered list.
@@ -11,9 +13,9 @@ namespace PHPForge\Html;
  */
 final class Ol extends Base\AbstractList
 {
-    use Attribute\Input\HasType;
-    use Attribute\Tag\HasReversed;
-    use Attribute\Tag\HasStart;
+    use HasReversed;
+    use HasStart;
+    use HasType;
 
     protected string $tagName = 'ol';
 }
