+++ src/Grouping/Ul.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PHPForge\Html;
+namespace PHPForge\Html\Grouping;
 
 /**
  * The `<ul>` HTML element represents an unordered list of items, typically rendered as a bulleted list.
