+++ tests/Grouping/Li/CustomMethodTest.php
@@ -2,10 +2,9 @@
 
 declare(strict_types=1);
 
-namespace PHPForge\Html\Tests\Li;
+namespace PHPForge\Html\Tests\Grouping\Li;
 
-use PHPForge\Html\Li;
-use PHPForge\Html\Ul;
+use PHPForge\Html\Grouping\{Li, Ul};
 use PHPForge\Support\Assert;
 use PHPUnit\Framework\TestCase;
